requirements.txt
torch>=2.0.0
customtkinter>=5.2.0
psutil>=5.9.0
numpy>=1.24.0
transformers>=4.35.0
safetensors>=0.4.1
python-telegram-bot==20.3